Sh Intalacion de requisitos 2.0

+++ install_requisites.sh
@@ -10,13 +10,14 @@ sudo apt update && sudo apt upgrade -y
 echo "Instalando Python 3 y pip..."
 sudo apt install -y python3 python3-pip
 
-# Instalar MySQL Server
-echo "Instalando MySQL Server..."
-sudo apt install -y mysql-server
-
-# Iniciar y habilitar MySQL
-echo "Habilitando e iniciando el servicio MySQL..."
-sudo systemctl enable --now mysql
+# Verificar si MySQL ya está instalado
+if dpkg -l | grep -q mysql-server; then
+    echo "MySQL ya está instalado. Omitiendo instalación."
+else
+    echo "MySQL no está instalado. Procediendo con la instalación..."
+    sudo apt install -y mysql-server
+    sudo systemctl enable --now mysql
+fi
 
 # Instalar conector de Python para MySQL
 echo "Instalando mysql-connector-python..."
